+++ src/QueryLexer/types.ts
@@ -1,4 +1,4 @@
-import type { Position } from '../common/types';
+import type { Position } from '../logic/types';
 import type { TokenContext } from '../QueryParser';
 import type { TokenTypeValues } from './logic/constants';
 
@@ -7,8 +7,14 @@ export type Token = {
   value: string;
   position: Position;
   context?: TokenContext;
+  prev: Token | null;
+  next: Token | null;
 };
 
 export type QueryLexerOptions = {
+  /**
+   * If true, the operators will be case sensitive.
+   * i.e. "AND", and "aNd" will be accepted on `false`, but not on `true`.
+   */
   caseSensitiveOperators?: boolean;
 };