+++ src/index.ts
@@ -1,13 +1,13 @@
-export * from './ASTBuilder';
+export * from './ASTUtils';
 export * from './QueryParser';
 export * from './QueryLexer';
 export * from './TokenStream';
 
-export type * from './ASTBuilder';
+export type * from './ASTUtils';
 export type * from './QueryParser';
 export type * from './QueryLexer';
 export type * from './TokenStream';
 
 // common
-export * from './common/constants';
-export type * from './common/constants';
+export * from './logic/constants';
+export type * from './logic/constants';